move the struct VectorFieldSettings to a new file
